requirements.txt
Django>=4.1,<4.1.5
django-cors-headers>=3.11.0,<3.14.0
django-rest-swagger>=2.2.0,<2.3.0
djangorestframework>=3.13.0,<3.15.0
psycopg2-binary==2.9.7
python-dotenv==0.21.0
drf-yasg==1.21.4
Pillow==10.0.1
gunicorn>=20.0.4,<20.2.0
boto3>=1.28,<1.28.50
django-storages>=1.12,<1.15